testdata: shuffle input files
The algorithm is quite sensitive to patterns (intentional or not) in the
input data.  In this case, "something" in data3 and data4 were causing a
number of estimators to get stuck.  This was visible by constructing a
very large number of estimators (~10000) and looking at the distribution
of the resulting estimates.  Shuffling the test data removes these
patterns and the estimators behaviour more sanely.  Note that this
property of the algorithm can not be fixed.